+++ tests/storybook/components/common/heading.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { faBug } from '@fortawesome/free-solid-svg-icons';
+import { faBug, faCog, faSync } from '@fortawesome/free-solid-svg-icons';
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import withPropsCombinations from 'react-storybook-addon-props-combinations';
@@ -17,7 +17,7 @@ storiesOf('Common', module)
         size:      [1, 2, 3, 4, 5, 6],
         icon:      [faBug],
         underline: [true, false],
-        controls:  [<div><Icon name="gear" /><Icon name="refresh" /></div>]
+        controls:  [<div><Icon name={faCog} /><Icon name={faSync} /></div>]
       },
       {
         CombinationRenderer({ Component, props }) {